SampleApp can be built with sdk source or sdk as static lib...
Added a second target to the SampleApp project called SampleAppSrc.  For
this target, the header search path is the src/include directory, while
the SampleApp target has a header search path of OpenKitSDK/include.

The purpose of this is to make it easy to test changes to the SDK
without building the static lib first and then including that in
the sample project.