Reintroduced OKCloudAsyncRequest to fix cloud storage bug, explained:
Wrapping all http interaction into a single class, OKNetworker,
is good.  At the moment, though, the setup surrounding the AFNetworking
HTTPClient is different for cloud storage versus leaderboards.

In OKCloud, it is necessary to explicitly set the http headers on
the NSMutableURLRequest returned from
-[AFHTTPClient-requestWithMethod:path:parameters].  See the comment
in the body of -[OKCloudAsyncRequest performWithCompletionHandler:].

Also notice, in OKCloudAsyncRequsts, we do not register the
AFJSONRequestOperation class, as we do for leaderboards.

These differences can of course be reconciled into a single class,
but to fix this bug immediately I am reverting back to the
OKCloudAsyncRequest class.