Not including parameters in HTTP POST in oauth signature:
By default, AFOAuth1Client includes POST params when computing request
signature, regardless of content-type.  Spec 1.0a states that POST
params should be included only for a content-type of
application/x-www-form-urlencoded.  At least that is how I interpreted
it.

With this mod in place, GETs and POSTs are both working properly with
the openkit backend.  Content-type is always application/json in POSTs
to our web services.